Fix nonce timestamp format mismatch
The current culture is used to create a string version of the timestamp,
but ParseTimestamp uses CultureInfo.InvariantCulture to parse it. If the
current culture differs from invariant the parsed timestamp will be wrong,
and the timestamp delta will be too large. This leads to never-ending
Digest challenges, with the password never accepted.

To reproduce the issue launch `Intl.cpl` and set the date format to
"Thai (Thailand)", then restart the web server.